Remove redundant static_cast<int>() for fields.count() and metaData.count().
loadFieldData() likely had the cast because originally, RomFields wasn't
a separate class; it was an std::vector<>

loadMetaData() was likely just a copy/paste thing.

Fixes several clang-tidy warnings:

warning: redundant explicit casting to the same type 'int' as the sub-expression,
remove this casting [readability-redundant-casting]

+++ src/libromdata/Audio/ADX.cpp
@@ -2,7 +2,7 @@
  * ROM Properties Page shell extension. (libromdata)                       *
  * ADX.hpp: CRI ADX audio reader.                                          *
  *                                                                         *
- * Copyright (c) 2018-2025 by David Korth.                                 *
+ * Copyright (c) 2018-2026 by David Korth.                                 *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
@@ -328,7 +328,7 @@ int ADX::loadFieldData(void)
 	}
 
 	// Finished reading the field data.
-	return static_cast<int>(d->fields.count());
+	return d->fields.count();
 }
 
 /**
@@ -369,7 +369,7 @@ int ADX::loadMetaData(void)
 		convSampleToMs(sample_count, sample_rate));
 
 	// Finished reading the metadata.
-	return static_cast<int>(d->metaData.count());
+	return d->metaData.count();
 }
 
 } // namespace LibRomData
