+++ src/libromdata/Console/GameCubeBNR.cpp
@@ -2,7 +2,7 @@
  * ROM Properties Page shell extension. (libromdata)                       *
  * GameCubeBNR.cpp: Nintendo GameCube banner reader.                       *
  *                                                                         *
- * Copyright (c) 2016-2024 by David Korth.                                 *
+ * Copyright (c) 2016-2026 by David Korth.                                 *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
@@ -591,7 +591,7 @@ int GameCubeBNR::loadFieldData(void)
 
 	if (d->comments.empty()) {
 		// Banner comment data wasn't loaded...
-		return static_cast<int>(d->fields.count());
+		return d->fields.count();
 	}
 	d->fields.reserve(3);	// Maximum of 3 fields.
 
@@ -719,7 +719,7 @@ int GameCubeBNR::loadFieldData(void)
 	}
 
 	// Finished reading the field data.
-	return static_cast<int>(d->fields.count());
+	return d->fields.count();
 }
 
 /**
@@ -796,7 +796,7 @@ int GameCubeBNR::loadMetaData(void)
 	d->metaData.addMetaData_string(Property::Description, d->getGameDescriptionString(comment));
 
 	// Finished reading the metadata.
-	return static_cast<int>(d->metaData.count());
+	return d->metaData.count();
 }
 
 /**
